Organized files, added fbaah7 and paulqasq parts, changes to productCard.css
