+++ .gitignore
@@ -1,3 +1,3 @@
 nbproject/
 test/
-no.html
+public_html/no.html
