--- public_html/bookstore-development/bookstore-development/Gruntfile.js
@@ -1,10 +1,0 @@
-/* 
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-module.exports = function (grunt) {
-    // Project configuration.
-    grunt.initConfig({
-    });
-};
