--- public_html/bookstore-development/bookstore-development/README.md
@@ -1,1 +1,0 @@
-# bookstore
