--- public_html/bookstore-development/bookstore-development/bower.json
@@ -1,13 +1,0 @@
-{
-    "name": "CS560Project",
-    "version": "1.0.0",
-    "main": "path/to/main.css",
-    "ignore": [
-        ".jshintrc",
-        "**/*.txt"
-    ],
-    "dependencies": {
-    },
-    "devDependencies": {
-    }
-}
