--- public_html/bookstore-development/bookstore-development/gulpfile.js
@@ -1,11 +1,0 @@
-/* 
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-
-var gulp = require('gulp');
-
-gulp.task('default', function () {
-    // place code for your default task here
-});
