--- public_html/bookstore-development/bookstore-development/package.json
@@ -1,8 +1,0 @@
-{
-    "name": "CS560Project",
-    "version": "1.0.0",
-    "keywords": ["util", "functional", "server", "client", "browser"],
-    "author": "tts-macbook",
-    "contributors": [],
-    "dependencies": {}
-}
