--- public_html/bookstore-development/bookstore-development/public_html/no.html
@@ -1,95 +1,0 @@
-<section class="container content-section" style="margin-top:30px">
-        <h2 class="section-header">Books</h2>
-        <div class="shop-books">
-            <div class="shop-book">
-                <span class="shop-book-title">The Great Gatsby</span>
-                <img class="shop-book-image" src="imgs/theGreatGatsby.jpg">
-                <div class="shop-book-details">
-                    <span class="Shop-book-price">$14.99</span>
-                    <button class="btn btn-primary shop-item-button" type="button">Add To Cart</button>
-                </div>
-            </div>
-            <div class="shop-book">
-                <span class="shop-book-title">The Great Gatsby</span>
-                <img class="shop-book-image" src="imgs/theGreatGatsby.jpg">
-                <div class="shop-book-details">
-                    <span class="Shop-book-price">$14.99</span>
-                    <button class="btn btn-primary shop-item-button" type="button">Add To Cart</button>
-                </div>
-            </div>
-            <div class="shop-book">
-                <span class="shop-book-title">The Great Gatsby</span>
-                <img class="shop-book-image" src="imgs/theGreatGatsby.jpg">
-                <div class="shop-book-details">
-                    <span class="Shop-book-price">$14.99</span>
-                    <button class="btn btn-primary shop-item-button" type="button">Add To Cart</button>
-                </div>
-            </div>
-        </div>
-    </section>
-
-
-
-
-
-<div class="card">
-            <img src="imgs/theGreatGatsby.jpg"  style="width:100%">
-            <h1>The Great Gatsby</h1>
-            <p class ="price">$14.99</p>
-            <p>The Great Gatsby is a novel by F. Scott Fitzgerald. It is a classic good novel.</p>
-            <p><button>Add to cart</button></p>
-        </div>
-
-
-
-<nav class="navbar navbar-expand-sm bg-dark navbar-dark">
-  <a class="navbar-brand" href="#">Nav</a>
-  <button class="navbar-toggler" type="button" data-toggle="collapse" data-target="#collapsibleNavbar">
-    <span class="navbar-toggler-icon"></span>
-  </button>
-  <div class="collapse navbar-collapse" id="collapsibleNavbar">
-    <ul class="navbar-nav">
-      <li class="nav-item">
-        <a class="nav-link" href="#">Home</a>
-      </li>
-      <li class="nav-item">
-        <a class="nav-link" href="#">Contact</a>
-      </li>
-      <li class="nav-item">
-        <a class="nav-link" href="#">Ummmm</a>
-      </li>
-    </ul>
-      <ul class ="nav navbar-nav navbar-right">
-          <div class="search-container">
-              <input type="text" id="myInput" onkeyup="searchFunction" placeholder="Search..." title="Search Book">
-          </div>
-          <li><a href="#"><span class="glyphicon glyphicon-shopping-cart"></span>Cart</a></li>
-      </ul>
-  </div> 
-</nav>
-
-    <nav class="navbar navbar-inverse">
-        <div class="container-fluid">
-            <div class="navbar-header">
-                <button type="button" class="navbar-toggle" data-toggle="collapse" data-target="#myNavbar">
-                    <span class="icon-bar"></span>
-                    <span class="icon-bar"></span>
-                    <span class="icon-bar"></span>
-                </button>
-                <a class="navbar-brand" href="#">CtBooks</a>
-            </div>
-            <div class="collapse navbar-collapse" id="myNavbar">
-                <ul class="nav navbar-nav">
-                    <li class="active"><a href="#">Home</a></li>
-                    <li><a href="#">About</a></li>
-                    <li><a href="#">Contact</a></li>
-                </ul>
-                <ul class="nav navbar-nav navbar-right">
-                    <li><a href="#"><span class="glyphicon glyphicon-shopping-cart"></span>Cart</a></li>
-                </ul>
-            </div>
-        </div>
-    </nav>
-
-
-<input type="text" id="myInput" onkeyup="searchFunction()" placeholder="Search...">
